requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.0.0
Werkzeug==3.1.4
gunicorn==23.0.0
